requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=11.0.3
python-multipart>=0.0.6
pydantic>=2.4.2
redis>=5.0.1
pymysql>=1.1.0
chromadb>=0.4.18
langchain>=0.1.0
python-dotenv>=1.0.0
celery>=5.3.6